Some fixes in A/P. GID is synced.
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@2356 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/dso-l2/src/com/tc/objectserver/persistence/impl/TransactionStoreImpl.java
@@ -20,14 +20,14 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Map;
-import java.util.SortedSet;
-import java.util.TreeSet;
+import java.util.SortedMap;
+import java.util.TreeMap;
 
 public class TransactionStoreImpl implements TransactionStore {
 
   private final Map                  serverTransactionIDMap = Collections.synchronizedMap(new HashMap());
-  private final SortedSet            ids                    = Collections
-                                                                .synchronizedSortedSet(new TreeSet(
+  private final SortedMap            ids                    = Collections
+                                                                .synchronizedSortedMap(new TreeMap(
                                                                                                    GlobalTransactionID.COMPARATOR));
   private final TransactionPersistor persistor;
   private final Sequence             globalIDSequence;
@@ -49,14 +49,20 @@ public class TransactionStoreImpl implements TransactionStore {
     gtx.commitComplete();
   }
 
+  // TODO:: This method maynot be needed anymore
   public GlobalTransactionDescriptor getTransactionDescriptor(ServerTransactionID serverTransactionID) {
     return (GlobalTransactionDescriptor) this.serverTransactionIDMap.get(serverTransactionID);
   }
 
-  public GlobalTransactionDescriptor createTransactionDescriptor(ServerTransactionID serverTransactionID) {
-    GlobalTransactionDescriptor rv = new GlobalTransactionDescriptor(serverTransactionID, getNextGlobalTransactionID());
-    basicAdd(rv);
-    return rv;
+  public GlobalTransactionDescriptor getOrCreateTransactionDescriptor(ServerTransactionID serverTransactionID) {
+    synchronized (serverTransactionIDMap) {
+      GlobalTransactionDescriptor rv = (GlobalTransactionDescriptor) serverTransactionIDMap.get(serverTransactionID);
+      if (rv == null) {
+        rv = new GlobalTransactionDescriptor(serverTransactionID, getNextGlobalTransactionID());
+        basicAdd(rv);
+      }
+      return rv;
+    }
   }
 
   private GlobalTransactionID getNextGlobalTransactionID() {
@@ -67,14 +73,14 @@ public class TransactionStoreImpl implements TransactionStore {
     ServerTransactionID sid = gtx.getServerTransactionID();
     GlobalTransactionID gid = gtx.getGlobalTransactionID();
     Object prevDesc = this.serverTransactionIDMap.put(sid, gtx);
-    ids.add(gid);
+    ids.put(gid, gtx);
     if (prevDesc != null) { throw new AssertionError("Adding new mapping for old txn IDs : " + gtx + " Prev desc = "
                                                      + prevDesc); }
   }
 
   public GlobalTransactionID getLeastGlobalTransactionID() {
     synchronized (ids) {
-      return (GlobalTransactionID) ((ids.isEmpty()) ? GlobalTransactionID.NULL_ID : ids.first());
+      return (GlobalTransactionID) ((ids.isEmpty()) ? GlobalTransactionID.NULL_ID : ids.firstKey());
     }
   }
 
@@ -90,7 +96,9 @@ public class TransactionStoreImpl implements TransactionStore {
         }
       }
     }
-    persistor.deleteAllByServerTransactionID(tx, toDelete);
+    if (!toDelete.isEmpty()) {
+      persistor.deleteAllByServerTransactionID(tx, toDelete);
+    }
   }
 
   public GlobalTransactionID getGlobalTransactionID(ServerTransactionID stxnID) {
@@ -114,4 +122,10 @@ public class TransactionStoreImpl implements TransactionStore {
     }
     removeAllByServerTransactionID(tx, stxIDs);
   }
+
+  public void createGlobalTransactionDesc(ServerTransactionID stxnID, GlobalTransactionID globalTransactionID) {
+    GlobalTransactionDescriptor rv = new GlobalTransactionDescriptor(stxnID, globalTransactionID);
+    basicAdd(rv);
+  }
+
 }
